+++ components/amorphic/lib/startPersistorMode.js
@@ -39,12 +39,14 @@ function startPersistorMode(appDirectory, logger, statsdClient, configStore = nu
             typeof logger.warn === 'function') {
 			SupertypeSession.logger.setLogger(logger);
         } else {
-			SupertypeSession.logger.warn({
+			SupertypeSession.logger.error({
 				module: moduleName,
 				function: functionName,
 				category: 'request',
-				message: 'sendToLog is deprecated, please pass in a logger here instead'
+				isHumanRelated: true,
+				message: 'no bunyan logger passed in'
 			});
+			throw new Error('sendToLog is deprecated, please pass in a bunyan logger');
 		}
 	}