+++ components/amorphic/HISTORY.md
@@ -1,5 +1,6 @@
 ## 11.0.0
 * Logger updates to use SupertypeLogger and the internal logger changes
+* changing sendToLog in listen and startPersistorMode to be logger instead for passing in a client logger to SupertypeLogger
 ## 10.3.0
 * Add validator middleware to express server, config settings for whitelist, blacklist, escape
 ## 10.2.1